Add local driving license application feature and supporting forms
Added functionality for users to create and edit local driving license applications, including selecting the intended license class.

Introduced a form to add or edit local driving license applications.

Added a management form to view and manage the list of local driving license applications.

Created a base class clsApplication to serve as a common parent for all application types in the future.